refactor: viewexam package 경로 수정
