--- usw-suwiki-post/usw-suwiki-viewexam/src/main/java/usw/suwiki/domain/user/viewexam/ViewExamRepository.java
@@ -1,14 +1,0 @@
-package usw.suwiki.domain.user.viewexam;
-
-import java.util.List;
-
-public interface ViewExamRepository {
-
-    void save(ViewExam viewExam);
-
-    List<ViewExam> findByUserId(Long userIdx);
-
-    void delete(ViewExam viewExam);
-
-    boolean validateIsExists(Long userId, Long lectureId);
-}
