+++ usw-suwiki-post/usw-suwiki-viewexam/src/main/java/usw/suwiki/domain/viewexam/ViewExamQueryRepository.java
@@ -1,9 +1,9 @@
-package usw.suwiki.domain.user.viewexam;
+package usw.suwiki.domain.viewexam;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import usw.suwiki.domain.user.viewexam.dto.ViewExamResponse;
+import usw.suwiki.domain.viewexam.dto.ViewExamResponse;
 
 import java.util.List;
