+++ usw-suwiki-post/usw-suwiki-viewexam/src/main/java/usw/suwiki/domain/viewexam/service/ClearViewExamServiceImpl.java
@@ -1,11 +1,11 @@
-package usw.suwiki.domain.user.viewexam.service;
+package usw.suwiki.domain.viewexam.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import usw.suwiki.domain.user.service.ClearViewExamService;
-import usw.suwiki.domain.user.viewexam.ViewExam;
-import usw.suwiki.domain.user.viewexam.ViewExamRepository;
+import usw.suwiki.domain.viewexam.ViewExam;
+import usw.suwiki.domain.viewexam.ViewExamRepository;
 
 @Service
 @Transactional
